:recycle: Orgnize constants and rename files
