+++ src/constants/app.ts
@@ -1,1 +1,16 @@
 export const SNACKBAR_AUTO_HIDE_DURATION = 6000;
+
+export const LOCAL_STORAGE_KEY = "skyrim-control-mapper-file";
+
+export const UPLOAD_BUTTON_TEXT = "Upload controlmap.txt";
+export const DOWNLOAD_BUTTON_TEXT = "Download controlmap.txt";
+
+export const DEFAULT_NEW_FORMAT_OPTION = {
+    label: "Skyrim AE 1.6.640 (New Format)",
+    filename: "Skyrim_AE_NewFormat",
+};
+
+export const DEFAULT_OLD_FORMAT_OPTION = {
+    label: "Skyrim SE 1.5.97 (Old Format)",
+    filename: "Skyrim_SE_OldFormat",
+};
